requirements.txt
pytest
Appium-Python-Client
allure-pytest
pytest-xdist